Implemented updated auth process [OGDEV-87]
