--- back-end/courier/src/main/java/com/jsg/courier/CourierApplication.java
@@ -1,17 +1,0 @@
-package com.jsg.courier;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-
-@SpringBootApplication
-@EnableAutoConfiguration
-@ComponentScan
-public class CourierApplication {
-	
-	public static void main(String[] args) {
-		SpringApplication.run(CourierApplication.class, args);
-		System.out.println("This application is now running...");
-	}
-}
