--- back-end/courier/src/main/java/com/jsg/courier/repositories/SQLRepository.java
@@ -1,14 +1,0 @@
-package com.jsg.courier.repositories;
-
-import java.util.List;
-
-public interface SQLRepository<T> {
-
-	public Boolean save(T item) throws Exception;
-	
-	public <V> List<T> findWhereEqual(String searchColumn, V value);
-	public <V> List<T> findWhereEqual(String searchColumn, V value, int limit);
-	
-	public Boolean closeConnection() throws Exception;
-	
-}
