--- back-end/courier/src/main/java/com/jsg/courier/utilities/AuthHeaderHandler.java
@@ -1,16 +1,0 @@
-package com.jsg.courier.utilities;
-
-public final class AuthHeaderHandler {
-	
-	public static String getBearerToken(String authHeader) {
-		if(authHeader == null) {
-			return null;
-		}
-		String[] headerParts = authHeader.split("Bearer");
-		if(headerParts.length != 2) {
-			return null;
-		}
-		return headerParts[1].trim();
-	}
-	
-}
