+++ back-end/courier_old/src/main/java/com/jsg/courier_old/api/websocket/SocketHandler.java
@@ -1,4 +1,4 @@
-package com.jsg.courier.api.websocket;
+package com.jsg.courier_old.api.websocket;
 
 import java.util.HashSet;
 import java.util.List;
@@ -6,8 +6,6 @@ import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.socket.CloseStatus;
 import org.springframework.web.socket.TextMessage;
@@ -15,29 +13,16 @@ import org.springframework.web.socket.WebSocketSession;
 import org.springframework.web.socket.handler.TextWebSocketHandler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.jsg.courier.datatypes.Message;
-import com.jsg.courier.datatypes.UserSession;
-import com.jsg.courier.datatypes.WebSocketHeaders;
-import com.jsg.courier.repositories.MessageRepository;
+import com.jsg.courier_old.datatypes.Message;
+import com.jsg.courier_old.datatypes.UserSession;
+import com.jsg.courier_old.datatypes.WebSocketHeaders;
+import com.jsg.courier_old.repositories.MessageRepository;
 
 @Service
 public class SocketHandler extends TextWebSocketHandler {
 		
 	private static ConcurrentHashMap<UUID, WebSocketSession> sessions = new ConcurrentHashMap<>();
 	private static final ObjectMapper objectMapper = new ObjectMapper();
-	
-	private final String SQL_CONNECTION_STRING;
-	private final String SQL_USERNAME;
-	private final String SQL_PASSWORD;
-	
-	@Autowired
-	public SocketHandler(@Value("${sql.username}") String sqlUsername,
-			@Value("${sql.password}") String sqlPassword,
-			@Value("${sql.connectionstring}") String sqlConnectionString) {
-		SQL_CONNECTION_STRING = sqlConnectionString;
-		SQL_USERNAME = sqlUsername;
-		SQL_PASSWORD = sqlPassword;
-	}
 		
 	@Override
 	public void handleTextMessage(WebSocketSession session, TextMessage messageJson) throws Exception {
@@ -99,10 +84,7 @@ public class SocketHandler extends TextWebSocketHandler {
 			if(i != 0) {
 				json += ",";
 			}
-			json += (new TextMessage(objectMapper
-					.writeValueAsString(new UserSession(headers.getId(),
-							SQL_CONNECTION_STRING, SQL_USERNAME, SQL_PASSWORD))))
-					.getPayload();
+			json += (new TextMessage(objectMapper.writeValueAsString(new UserSession(headers.getId())))).getPayload();
 			i++;
 		}
 		json += "]";
