+++ back-end/courier_old/src/main/java/com/jsg/courier_old/api/websocket/WebSocketHandshakeInterceptor.java
@@ -1,12 +1,10 @@
-package com.jsg.courier.api.websocket;
+package com.jsg.courier_old.api.websocket;
 
 import java.util.Map;
 
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.server.ServerHttpRequest;
 import org.springframework.http.server.ServerHttpResponse;
@@ -14,34 +12,19 @@ import org.springframework.http.server.ServletServerHttpRequest;
 import org.springframework.stereotype.Component;
 import org.springframework.web.socket.WebSocketHandler;
 import org.springframework.web.socket.server.HandshakeInterceptor;
-import org.springframework.web.util.WebUtils;
 
-import com.jsg.courier.utilities.JWTHandler;
+import com.jsg.courier_old.utilities.JWTHandler;
 
 @Component
 public class WebSocketHandshakeInterceptor implements HandshakeInterceptor {
-	
-	private static final String ACCESS_TOKEN_NAME = "acc.tok";
-	private final String ACCESS_TOKEN_SECRET;
-	
-	@Autowired
-	public WebSocketHandshakeInterceptor(@Value("${token.secret.access}") String accessTokenSecret) {
-		ACCESS_TOKEN_SECRET = accessTokenSecret;
-	}
-	
+
 	@Override
 	public boolean beforeHandshake(ServerHttpRequest request, ServerHttpResponse response, WebSocketHandler wsHandler,
 			Map<String, Object> attributes) throws Exception {
 		HttpServletRequest req = ((ServletServerHttpRequest) request).getServletRequest();
-		
-		Cookie cookie = WebUtils.getCookie(req, ACCESS_TOKEN_NAME);
-		
-//		Cookie[] cookies = req.getCookies();
+		Cookie[] cookies = req.getCookies();
 		String protocolHeader = request.getHeaders().getFirst("sec-websocket-protocol");
-		System.out.println("Protocol header is:" + protocolHeader);
-		if(cookie == null || protocolHeader == null || protocolHeader.contentEquals("")) {
-			System.out.println("Cookies not present...");
-			System.out.println(cookie);
+		if(cookies == null || protocolHeader == null || protocolHeader.contentEquals("")) {
 			response.setStatusCode(HttpStatus.UNAUTHORIZED);
 			return false;
 		}
@@ -55,11 +38,10 @@ public class WebSocketHandshakeInterceptor implements HandshakeInterceptor {
 			response.setStatusCode(HttpStatus.UNAUTHORIZED);
 			return false;
 		}
-		String token = cookie.getValue();
-		System.out.println("Cookie token is: " + token);
+		String token = cookies[0].getValue();
 		String headerToken = protocolHeaders[1].trim();
-		if(!JWTHandler.tokenIsValid(token, ACCESS_TOKEN_SECRET) || 
-				!JWTHandler.tokenIsValid(headerToken, ACCESS_TOKEN_SECRET)) {
+		long id = Long.parseLong(idString);
+		if(!JWTHandler.tokenIsValid(token, id) || !JWTHandler.tokenIsValid(headerToken, id)) {
 			response.setStatusCode(HttpStatus.UNAUTHORIZED);
 			return false;
 		}
@@ -69,6 +51,7 @@ public class WebSocketHandshakeInterceptor implements HandshakeInterceptor {
 
 	@Override
 	public void afterHandshake(ServerHttpRequest request, ServerHttpResponse response, WebSocketHandler wsHandler,
-			Exception exception) {}
+			Exception exception) {
+	}
 
 }
