+++ back-end/courier_old/src/main/java/com/jsg/courier_old/datatypes/User.java
@@ -1,0 +1,49 @@
+package com.jsg.courier_old.datatypes;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public class User {
+	
+	@JsonProperty
+	private long id;
+	
+	@JsonProperty
+	private String email;
+	
+	@JsonProperty
+	private String password;
+	
+	@JsonProperty
+	private String salt;
+	
+	public User() {}
+	
+	public User(String email, String password) {
+		this.email = email;
+		this.password = password;
+	}
+	
+	public User(String email, String password, long id) {
+		this.email = email;
+		this.password = password;
+		this.id = id;
+	}
+	
+	public String getEmail() {
+		return this.email;
+	}
+	
+	public long getId() {
+		return this.id;
+	}
+	
+	public String getPassword() {
+		return this.password;
+	}
+	
+	public void clearPassword() {
+		this.password = null;
+		this.salt = null;
+	}
+	
+}
