+++ back-end/courier_old/src/main/java/com/jsg/courier_old/datatypes/UserInfo.java
@@ -1,0 +1,42 @@
+package com.jsg.courier_old.datatypes;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public class UserInfo {
+
+	@JsonProperty
+	private long id;
+	
+	@JsonProperty
+	private String displayName;
+	
+	@JsonProperty
+	private String bio;
+	
+	public UserInfo() {}
+	
+	public UserInfo(long id, String name) {
+		this.id = id;
+		this.displayName = name;
+		this.bio = "";
+	}
+	
+	public UserInfo(long id, String displayName, String bio) {
+		this.id = id;
+		this.displayName = displayName;
+		this.bio = bio;
+	}
+	
+	public long getId() {
+		return this.id;
+	}
+	
+	public String getDisplayName() {
+		return this.displayName;
+	}
+	
+	public String getBio() {
+		return this.bio;
+	}
+	
+}
