+++ back-end/courier_old/src/main/java/com/jsg/courier_old/repositories/SQLRepository.java
@@ -1,0 +1,14 @@
+package com.jsg.courier_old.repositories;
+
+import java.util.List;
+
+public interface SQLRepository<T> {
+
+	public Boolean save(T item) throws Exception;
+	
+	public <V> List<T> findWhereEqual(String searchColumn, V value);
+	public <V> List<T> findWhereEqual(String searchColumn, V value, int limit);
+	
+	public Boolean closeConnection() throws Exception;
+	
+}
