+++ back-end/courier_old/src/main/java/com/jsg/courier_old/utilities/AuthHeaderHandler.java
@@ -1,0 +1,16 @@
+package com.jsg.courier_old.utilities;
+
+public final class AuthHeaderHandler {
+	
+	public static String getBearerToken(String authHeader) {
+		if(authHeader == null) {
+			return null;
+		}
+		String[] headerParts = authHeader.split("Bearer");
+		if(headerParts.length != 2) {
+			return null;
+		}
+		return headerParts[1].trim();
+	}
+	
+}
